Separate webhook server from operator
